+++ src-docs/src/views/combo_box/combo_box.js
@@ -77,6 +77,7 @@ export default class extends Component {
         onChange={this.onChange}
         onCreateOption={this.onCreateOption}
         isClearable={true}
+        data-test-subj="demoComboBox"
       />
     );
   }