+++ src-docs/src/views/xy_chart/empty.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import { EuiXYChart } from '../../../../src/experimental';
-
-export default () => <EuiXYChart width={600} height={200} statusText="Missing data"/>;