+++ src-docs/src/routes.js
@@ -214,22 +214,22 @@ import { ToggleExample }
   from './views/toggle/toggle_example';
 
 import { XYChartExample }
-  from './views/xy_chart/xy_chart_example';
+  from './views/series_chart/series_chart_example';
 
 import { XYChartAxisExample }
-  from './views/xy_chart_axis/xy_axis_example';
+  from './views/series_chart_axis/series_axis_example';
 
 import { XYChartBarExample }
-  from './views/xy_chart_bar/bar_example';
+  from './views/series_chart_bar/bar_example';
 
 import { XYChartHistogramExample }
-  from './views/xy_chart_histogram/histogram_example';
+  from './views/series_chart_histogram/histogram_example';
 
 import { XYChartAreaExample }
-  from './views/xy_chart_area/area_example';
+  from './views/series_chart_area/area_example';
 
 import { XYChartLineExample }
-  from './views/xy_chart_line/line_example';
+  from './views/series_chart_line/line_example';
 
 import { Changelog }
   from './views/package/changelog';
@@ -363,7 +363,7 @@ const navigation = [{
   ].map(example => createExample(example)),
 },
 {
-  name: 'XY Charts (Beta)',
+  name: 'Series charts (beta)',
   items: [
     XYChartExample,
     XYChartAxisExample,