+++ src/components/series_chart/utils/index.js
@@ -1,0 +1,3 @@
+export { EuiSeriesChartUtils } from './chart_utils';
+export { EuiSeriesChartAxisUtils } from './axis_utils';
+export { EuiSeriesChartTextUtils } from './text_utils';