+++ src/components/xy_chart/utils/index.js
@@ -1,3 +1,0 @@
-export { EuiXYChartUtils } from './chart_utils';
-export { EuiXYChartAxisUtils } from './axis_utils';
-export { EuiXYChartTextUtils } from './text_utils';