+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fixed `EuiXYChart` responsive resize in a flexbox layout ([#1041](https://github.com/elastic/eui/pull/1041))
 
+**Experimental breaking change**
+
+ - Renamed `EuiXYChart` to `EuiSeriesChart`, `EuiXYChartUtils` to `EuiSeriesChartUtils`, `EuiXYChartAxisUtils` to `EuiSeriesChartAxisUtils`, and  `EuiXYChartTextUtils` to `EuiSeriesChartTextUtils` ([#](https://github.com/elastic/eui/pull/))
+
 ## [`3.2.1`](https://github.com/elastic/eui/tree/v3.2.1)
 
 - Added `closeButtonAriaLabel` property to `EuiFlyout` ([#1031](https://github.com/elastic/eui/pull/1031))