+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
-No public interface changes since `3.3.0`.
+**Bug fixes**
+
+- `EuiMutationObserver`'s `children` prop is no longer marked as required ([#1076](https://github.com/elastic/eui/pull/1076))
 
 ## [`3.3.0`](https://github.com/elastic/eui/tree/v3.3.0)